Update code TLG

+++ src/Common/ISD.Infrastructure/ISD.Infrastructure/Models/ProductModel.cs
@@ -15,7 +15,7 @@ namespace ISD.Infrastructure.Models
         public Guid ProductId { get; set; }
         [StringLength(50)]
         public string ProductCode { get; set; }
-        public int? ProductCodeInt { get; set; }
+        public long? ProductCodeInt { get; set; }
         [StringLength(50)]
         public string ProductName { get; set; }
         [StringLength(50)]
